+++ src/common/mux.h
@@ -1,5 +1,5 @@
 #pragma once
 
 
-int mux_sample(int addr);
+unsigned short mux_sample(int addr);
 bool get_hall(int addr);
